CLI-837 Extract declarative integration replay into core/framework/features/reconcile.ts
post-update.ts's migrateDeclarativeIntegrations needed findInstalledIntegration,
integrationInstaller, isFeatureContainer, and FeatureDeclaration from
core/framework/features, making core/host depend on core/framework while
framework already depends on host for its own installers. Move the
replay/reconciliation logic into framework as reconcileInstalledIntegrations,
which framework's applyAndRecordFeatures already models via its
executionMode: 'update' concept. post-update.ts now only needs the
IntegrationRegistry type plus this one function.
