+++ src/core/host/post-update.ts
@@ -23,13 +23,8 @@ import { homedir } from 'node:os';
 import { join } from 'node:path';
 
 import {
-  type FeatureApplication,
-  type FeatureDeclaration,
-  findInstalledIntegration,
-  type IntegrationDeclaration,
-  integrationInstaller,
   type IntegrationRegistry,
-  isFeatureContainer,
+  reconcileInstalledIntegrations,
 } from '@/core/framework/features';
 import { installScaScannerBinary } from '@/core/host/install/sca-scanner.ts';
 import { installSecretsBinary } from '@/core/host/install/secrets.ts';
@@ -38,13 +33,7 @@ import { appendTelemetryEvent } from '@/core/telemetry/telemetry-events.ts';
 import { version as CURRENT_VERSION } from '../../../package.json';
 import { getTelemetryDir } from '../config-constants.ts';
 import logger from '../observability/logger.ts';
-import type {
-  CliState,
-  HookExtension,
-  InstalledIntegration,
-  InstalledIntegrationFeature,
-  IntegrationStateAttribute,
-} from '../state/state.ts';
+import type { CliState, HookExtension } from '../state/state.ts';
 import { loadState, saveState, stateFileExists, tryLoadState } from '../state/state-repository.ts';
 import { isNewerVersion } from '../version.ts';
 import { SCA_SCANNER_BINARY_NAME, SECRETS_BINARY_NAME } from './install-types.ts';
@@ -176,239 +165,9 @@ function migrateSinkFile(): void {
 
 export async function migrateDeclarativeIntegrations(registry: IntegrationRegistry): Promise<void> {
   const state = loadState();
-  let stateChanged = false;
-
-  for (const integration of registry.list()) {
-    const installedIntegration = findInstalledIntegration(state, integration);
-    if (!installedIntegration) {
-      continue;
-    }
-    const originalFeatures = [...installedIntegration.features];
-    const { applications, removedUnknownFeatures } = createMigrationFeatureApplications(
-      integration,
-      installedIntegration,
-    );
-    if (removedUnknownFeatures) {
-      stateChanged = true;
-    }
-
-    try {
-      const installedFeatures = await integrationInstaller.applyAndRecordFeatures(
-        state,
-        integration,
-        applications,
-        {
-          continueOnFeatureError: true,
-          executionMode: 'update',
-          onFeatureError: (application, err) => {
-            logger.debug(
-              `Declarative migration failed for ${integration.id}.${application.feature.id}: ${err.message}`,
-            );
-          },
-        },
-      );
-      if (installedFeatures.length > 0) {
-        stateChanged = true;
-      }
-    } catch (err) {
-      logger.debug(`Declarative migration failed for ${integration.id}: ${(err as Error).message}`);
-    }
-    restoreFailedReplacements(installedIntegration, originalFeatures, applications);
-  }
-
-  if (stateChanged) {
+  if (await reconcileInstalledIntegrations(state, registry)) {
     saveState(state);
   }
-}
-
-function createMigrationFeatureApplications(
-  integration: IntegrationDeclaration,
-  installedIntegration: InstalledIntegration,
-): { applications: FeatureApplication[]; removedUnknownFeatures: boolean } {
-  const featuresById = new Map(integration.features.map((feature) => [feature.id, feature]));
-  const replacementApplications = migrateReplacedFeatures(
-    integration,
-    featuresById,
-    installedIntegration.features,
-  );
-  const knownFeatures = installedIntegration.features.filter((feature) =>
-    featuresById.has(feature.featureId),
-  );
-  const removedUnknownFeatures = knownFeatures.length !== installedIntegration.features.length;
-  if (removedUnknownFeatures) {
-    installedIntegration.features = knownFeatures;
-  }
-
-  const applications: FeatureApplication[] = [];
-  for (const installedFeature of knownFeatures) {
-    const application = createFeatureApplication(
-      featuresById,
-      installedFeature.featureId,
-      installedFeature.subfeatures?.map((subfeature) => subfeature.featureId),
-      installedFeature.targetRoot,
-      installedFeature.scope,
-      installedFeature.attrs,
-    );
-    if (application) {
-      applications.push(application);
-    }
-  }
-  applications.push(...replacementApplications);
-
-  return { applications, removedUnknownFeatures };
-}
-
-function migrateReplacedFeatures(
-  integration: IntegrationDeclaration,
-  featuresById: Map<string, FeatureDeclaration>,
-  installedFeatures: InstalledIntegrationFeature[],
-): FeatureApplication[] {
-  const applications: FeatureApplication[] = [];
-
-  for (const successor of integration.features) {
-    const predecessorsByTarget = groupReplacedFeaturesByTarget(successor, installedFeatures);
-
-    for (const predecessors of predecessorsByTarget.values()) {
-      const { scope, targetRoot } = predecessors[0];
-      const hasRecordedSuccessor = installedFeatures.some(
-        (feature) =>
-          feature.featureId === successor.id &&
-          feature.scope === scope &&
-          feature.targetRoot === targetRoot,
-      );
-      if (hasRecordedSuccessor) {
-        continue;
-      }
-      const application = createFeatureApplication(
-        featuresById,
-        successor.id,
-        undefined,
-        targetRoot,
-        scope,
-        mergeFeatureAttrs(predecessors),
-      );
-      if (application) {
-        applications.push(application);
-      }
-    }
-  }
-
-  return applications;
-}
-
-/**
- * A state-only rollback may retain predecessor feature ids that no longer exist
- * in the current declarations. Post-update will retry the migration on the next
- * version update; a future command (for example, `sonar doctor`) should also
- * allow users to trigger these migrations without waiting for another release.
- */
-function restoreFailedReplacements(
-  installedIntegration: InstalledIntegration,
-  originalFeatures: InstalledIntegrationFeature[],
-  applications: FeatureApplication[],
-): void {
-  for (const application of applications) {
-    const replacedIds = application.feature.replacedIds;
-    if (!replacedIds?.length) {
-      continue;
-    }
-    const successorWasInstalled = installedIntegration.features.some(
-      (feature) =>
-        feature.featureId === application.feature.id &&
-        feature.scope === application.scope &&
-        feature.targetRoot === application.targetRoot,
-    );
-    if (successorWasInstalled) {
-      continue;
-    }
-
-    const predecessorsToRestore = originalFeatures.filter(
-      (originalFeature) =>
-        replacedIds.includes(originalFeature.featureId) &&
-        originalFeature.scope === application.scope &&
-        originalFeature.targetRoot === application.targetRoot &&
-        !installedIntegration.features.includes(originalFeature),
-    );
-    installedIntegration.features.push(...predecessorsToRestore);
-  }
-}
-
-function groupReplacedFeaturesByTarget(
-  successor: FeatureDeclaration,
-  installedFeatures: InstalledIntegrationFeature[],
-): Map<string, InstalledIntegrationFeature[]> {
-  const predecessorsByTarget = new Map<string, InstalledIntegrationFeature[]>();
-
-  for (const replacedId of successor.replacedIds ?? []) {
-    for (const installedFeature of installedFeatures) {
-      if (installedFeature.featureId !== replacedId) {
-        continue;
-      }
-      const targetKey = `${installedFeature.scope}:${installedFeature.targetRoot}`;
-      const predecessors = predecessorsByTarget.get(targetKey);
-      if (predecessors) {
-        predecessors.push(installedFeature);
-      } else {
-        predecessorsByTarget.set(targetKey, [installedFeature]);
-      }
-    }
-  }
-  return predecessorsByTarget;
-}
-
-function mergeFeatureAttrs(
-  features: InstalledIntegrationFeature[],
-): Record<string, IntegrationStateAttribute> | undefined {
-  const attrs = features.reduce<Record<string, IntegrationStateAttribute>>(
-    (merged, feature) => ({ ...merged, ...feature.attrs }),
-    {},
-  );
-  return Object.keys(attrs).length > 0 ? attrs : undefined;
-}
-
-function getFeature(
-  featuresById: Map<string, FeatureDeclaration>,
-  featureId: string,
-  subfeatureIds: string[] | undefined,
-): FeatureDeclaration | undefined {
-  const feature = featuresById.get(featureId);
-  if (!feature) {
-    return undefined;
-  }
-
-  let applicationFeature = feature;
-  if (isFeatureContainer(feature)) {
-    const activeIds = new Set(subfeatureIds ?? feature.defaultInstallSubfeatureIds);
-    const filteredContainer = {
-      ...feature,
-      subfeatures: feature.subfeatures.filter((s) => activeIds.has(s.id)),
-    };
-    applicationFeature = filteredContainer;
-  }
-  return applicationFeature;
-}
-
-function createFeatureApplication(
-  featuresById: Map<string, FeatureDeclaration>,
-  featureId: string,
-  subfeatureIds: string[] | undefined,
-  targetRoot: string,
-  scope: InstalledIntegrationFeature['scope'],
-  attrs: InstalledIntegrationFeature['attrs'],
-): FeatureApplication | undefined {
-  const feature = getFeature(featuresById, featureId, subfeatureIds);
-  if (!feature) {
-    return undefined;
-  }
-
-  if (scope === 'project' && !fs.existsSync(targetRoot)) {
-    logger.debug(
-      `Declarative migration skipped for ${featureId}: target root no longer exists: ${targetRoot}`,
-    );
-    return undefined;
-  }
-
-  return { feature, targetRoot, scope, attrs };
 }
 
 /**
