+++ tests/unit/core/host/post-update.test.ts
@@ -29,14 +29,6 @@ import { supportedIntegrations } from '@/commands/integrate';
 import { CLAUDE_INTEGRATION_ID } from '@/commands/integrate/claude/declaration.ts';
 import * as hooks from '@/commands/integrate/claude/hooks.ts';
 import * as configConstants from '@/core/config-constants.ts';
-import {
-  type ContainerIntegrationContext,
-  type DependencyDeclaration,
-  type FeatureContainer,
-  type IntegrationContext,
-  IntegrationRegistry,
-  wholeFile,
-} from '@/core/framework/features';
 import { DISTRIBUTION } from '@/core/host/distribution.ts';
 import * as scaScannerInstall from '@/core/host/install/sca-scanner.ts';
 import * as secretsInstall from '@/core/host/install/secrets.ts';
@@ -44,19 +36,13 @@ import { SCA_SCANNER_BINARY_NAME } from '@/core/host/install-types.ts';
 import * as migration from '@/core/host/migration.ts';
 import {
   migrateClaudeCodeHooks,
-  migrateDeclarativeIntegrations,
   migrateLegacyTelemetryEvents,
   type PostUpdateDependencies,
   runPostUpdateActions,
   updateScaScannerBinaryIfNeeded,
   updateSecretsBinaryIfNeeded,
 } from '@/core/host/post-update.ts';
-import type {
-  CliState,
-  HookExtension,
-  InstalledIntegrationFeature,
-  StoredCommandExecutedEvent,
-} from '@/core/state/state.ts';
+import type { CliState, HookExtension, StoredCommandExecutedEvent } from '@/core/state/state.ts';
 import { getDefaultState } from '@/core/state/state.ts';
 import * as stateRepository from '@/core/state/state-repository.ts';
 import * as telemetryEvents from '@/core/telemetry/telemetry-events.ts';
@@ -336,684 +322,6 @@ describe('migrateLegacyTelemetryEvents', () => {
 
     expect(fs.existsSync(oldPath)).toBe(false);
     expect(fs.readFileSync(newPath, 'utf-8')).toBe('{"event":"existing"}\n{"event":"buffered"}\n');
-  });
-});
-
-describe('migrateDeclarativeIntegrations', () => {
-  let loadStateSpy: Mock<typeof stateRepository.loadState>;
-  let saveStateSpy: Mock<typeof stateRepository.saveState>;
-  let tempDir: string;
-
-  beforeEach(() => {
-    tempDir = fs.mkdtempSync(join(tmpdir(), 'sonar-cli-post-update-'));
-    loadStateSpy = spyOn(stateRepository, 'loadState').mockReturnValue(makeState());
-    saveStateSpy = spyOn(stateRepository, 'saveState').mockImplementation(() => {});
-  });
-
-  afterEach(() => {
-    loadStateSpy.mockRestore();
-    saveStateSpy.mockRestore();
-    fs.rmSync(tempDir, { recursive: true, force: true });
-  });
-
-  function recordedFeature(
-    featureId: string,
-    attrs?: InstalledIntegrationFeature['attrs'],
-  ): InstalledIntegrationFeature {
-    const now = '2026-01-01T00:00:00.000Z';
-    return {
-      featureId,
-      scope: 'project',
-      targetRoot: tempDir,
-      installedByCliVersion: '0.9.0',
-      installedAt: now,
-      updatedByCliVersion: '0.9.0',
-      updatedAt: now,
-      dependencies: [],
-      resources: [],
-      operations: [],
-      attrs,
-    };
-  }
-
-  it('reapplies only installed declarative features and prunes unknown feature state', async () => {
-    const operationCalls: string[] = [];
-    const resourcePath = join(tempDir, 'managed.txt');
-    const newFeatureResourcePath = join(tempDir, 'new-feature.txt');
-    const now = '2026-01-01T00:00:00.000Z';
-    fs.writeFileSync(resourcePath, 'legacy content', 'utf-8');
-
-    const state = makeState();
-    state.integrations.installed.push({
-      id: 'integration-id',
-      integrationId: 'test-integration',
-      installedByCliVersion: '0.9.0',
-      installedAt: now,
-      updatedByCliVersion: '0.9.0',
-      updatedAt: now,
-      features: [
-        {
-          featureId: 'managed-feature',
-          scope: 'project',
-          targetRoot: tempDir,
-          installedByCliVersion: '0.9.0',
-          installedAt: now,
-          updatedByCliVersion: '0.9.0',
-          updatedAt: now,
-          dependencies: [],
-          resources: [
-            {
-              id: 'managed-file',
-              resourceType: 'whole-file',
-              version: '1',
-              path: resourcePath,
-              updatedByCliVersion: '0.9.0',
-              updatedAt: now,
-            },
-          ],
-          operations: [],
-          attrs: { projectKey: 'project-key' },
-        },
-        {
-          featureId: 'removed-feature',
-          scope: 'project',
-          targetRoot: tempDir,
-          installedByCliVersion: '0.9.0',
-          installedAt: now,
-          updatedByCliVersion: '0.9.0',
-          updatedAt: now,
-          dependencies: [],
-          resources: [],
-          operations: [],
-        },
-      ],
-    });
-    loadStateSpy.mockReturnValue(state);
-
-    const registry = new IntegrationRegistry();
-    registry.register({
-      id: 'test-integration',
-      displayName: 'Test integration',
-      features: [
-        {
-          id: 'managed-feature',
-          displayName: 'Managed feature',
-          resources: [
-            wholeFile({
-              id: 'managed-file',
-              version: '2',
-              targetPath: resourcePath,
-              content: 'fresh content',
-            }),
-          ],
-          operations: [
-            {
-              id: 'refresh-operation',
-              version: '1',
-              apply: () => {
-                operationCalls.push('refresh-operation');
-              },
-            },
-          ],
-        },
-        {
-          id: 'new-feature',
-          displayName: 'New feature',
-          resources: [
-            wholeFile({
-              id: 'new-managed-file',
-              version: '1',
-              targetPath: newFeatureResourcePath,
-              content: 'should not be written automatically',
-            }),
-          ],
-          operations: [
-            {
-              id: 'new-feature-operation',
-              version: '1',
-              apply: () => {
-                operationCalls.push('new-feature-operation');
-              },
-            },
-          ],
-        },
-      ],
-    });
-
-    await migrateDeclarativeIntegrations(registry);
-
-    expect(fs.readFileSync(resourcePath, 'utf-8')).toBe('fresh content');
-    expect(operationCalls).toEqual(['refresh-operation']);
-    expect(fs.existsSync(newFeatureResourcePath)).toBe(false);
-    expect(saveStateSpy).toHaveBeenCalledTimes(1);
-
-    const savedState = saveStateSpy.mock.calls[0][0];
-    expect(savedState.integrations.installed).toHaveLength(1);
-    expect(savedState.integrations.installed[0].features).toHaveLength(1);
-    const savedFeature = savedState.integrations.installed[0].features[0];
-    expect(savedFeature.featureId).toBe('managed-feature');
-    expect(savedFeature.resources).toHaveLength(1);
-    expect(savedFeature.resources[0]).toMatchObject({
-      id: 'managed-file',
-      version: '2',
-      path: resourcePath,
-    });
-    expect(savedFeature.operations).toHaveLength(1);
-    expect(savedFeature.operations[0]).toMatchObject({
-      id: 'refresh-operation',
-      version: '1',
-    });
-  });
-
-  it('installs a shared dependency once and exposes the refreshed dependency to feature resources', async () => {
-    const dependencyPathA = join(tempDir, 'feature-a.txt');
-    const dependencyPathB = join(tempDir, 'feature-b.txt');
-    const existingDependencyPaths: string[] = [];
-    const installCalls: string[] = [];
-    const now = '2026-01-01T00:00:00.000Z';
-
-    const sharedDependency: DependencyDeclaration = {
-      id: 'shared-dependency',
-      dependencyType: 'binary',
-      version: '2',
-      installOrUpdate: ({ existingDependency }) => {
-        existingDependencyPaths.push(existingDependency?.path ?? 'missing');
-        installCalls.push('install');
-        return {
-          id: 'shared-dependency',
-          dependencyType: 'binary',
-          version: '2',
-          path: '/new/shared-dependency',
-        };
-      },
-      isInstalled: () => true,
-      remove: () => {},
-    };
-
-    const state = makeState();
-    state.integrations.installed.push({
-      id: 'integration-id',
-      integrationId: 'test-integration',
-      installedByCliVersion: '0.9.0',
-      installedAt: now,
-      updatedByCliVersion: '0.9.0',
-      updatedAt: now,
-      features: [
-        {
-          featureId: 'feature-a',
-          scope: 'project',
-          targetRoot: tempDir,
-          installedByCliVersion: '0.9.0',
-          installedAt: now,
-          updatedByCliVersion: '0.9.0',
-          updatedAt: now,
-          dependencies: [{ id: sharedDependency.id }],
-          resources: [],
-          operations: [],
-        },
-        {
-          featureId: 'feature-b',
-          scope: 'project',
-          targetRoot: tempDir,
-          installedByCliVersion: '0.9.0',
-          installedAt: now,
-          updatedByCliVersion: '0.9.0',
-          updatedAt: now,
-          dependencies: [{ id: sharedDependency.id }],
-          resources: [],
-          operations: [],
-        },
-      ],
-    });
-    state.dependencies.installed.push({
-      id: sharedDependency.id,
-      dependencyType: sharedDependency.dependencyType,
-      version: '1',
-      path: '/old/shared-dependency',
-      updatedByCliVersion: '0.9.0',
-      updatedAt: now,
-    });
-    loadStateSpy.mockReturnValue(state);
-
-    const registry = new IntegrationRegistry();
-    registry.register({
-      id: 'test-integration',
-      displayName: 'Test integration',
-      features: [
-        {
-          id: 'feature-a',
-          displayName: 'Feature A',
-          dependencies: [sharedDependency],
-          resources: [
-            wholeFile({
-              id: 'feature-a-file',
-              version: '1',
-              targetPath: dependencyPathA,
-              content: (context: IntegrationContext) =>
-                String(context.resolvedDependencies.get(sharedDependency.id)?.path),
-            }),
-          ],
-        },
-        {
-          id: 'feature-b',
-          displayName: 'Feature B',
-          dependencies: [sharedDependency],
-          resources: [
-            wholeFile({
-              id: 'feature-b-file',
-              version: '1',
-              targetPath: dependencyPathB,
-              content: (context: IntegrationContext) =>
-                String(context.resolvedDependencies.get(sharedDependency.id)?.path),
-            }),
-          ],
-        },
-      ],
-    });
-
-    await migrateDeclarativeIntegrations(registry);
-
-    expect(existingDependencyPaths).toEqual(['/old/shared-dependency']);
-    expect(installCalls).toEqual(['install']);
-    expect(fs.readFileSync(dependencyPathA, 'utf-8')).toBe('/new/shared-dependency');
-    expect(fs.readFileSync(dependencyPathB, 'utf-8')).toBe('/new/shared-dependency');
-    expect(saveStateSpy).toHaveBeenCalledTimes(1);
-  });
-
-  it('activates only migrationDefaultSubfeatureIds when upgrading from old plain-feature state', async () => {
-    const capturedContexts: IntegrationContext[] = [];
-    const now = '2026-01-01T00:00:00.000Z';
-
-    const state = makeState();
-    state.integrations.installed.push({
-      id: 'integration-id',
-      integrationId: 'test-integration',
-      installedByCliVersion: '0.9.0',
-      installedAt: now,
-      updatedByCliVersion: '0.9.0',
-      updatedAt: now,
-      features: [
-        {
-          featureId: 'container-feature',
-          scope: 'project',
-          targetRoot: tempDir,
-          installedByCliVersion: '0.9.0',
-          installedAt: now,
-          updatedByCliVersion: '0.9.0',
-          updatedAt: now,
-          dependencies: [],
-          resources: [],
-          operations: [],
-          // no subfeatures — simulates an old plain-feature install
-        },
-      ],
-    });
-    loadStateSpy.mockReturnValue(state);
-
-    const container: FeatureContainer = {
-      id: 'container-feature',
-      displayName: 'Container feature',
-      defaultInstallSubfeatureIds: ['sub-a'],
-      subfeatures: [
-        { id: 'sub-a', displayName: 'Sub A' },
-        { id: 'sub-b', displayName: 'Sub B' },
-      ],
-      operations: [
-        {
-          id: 'test-op',
-          apply: (ctx) => {
-            capturedContexts.push(ctx);
-          },
-        },
-      ],
-    };
-    const registry = new IntegrationRegistry();
-    registry.register({
-      id: 'test-integration',
-      displayName: 'Test integration',
-      features: [container],
-    });
-
-    await migrateDeclarativeIntegrations(registry);
-
-    expect(capturedContexts).toHaveLength(1);
-    expect('activeSubfeatures' in capturedContexts[0]).toBeTrue();
-    expect(
-      (capturedContexts[0] as ContainerIntegrationContext).activeSubfeatures.map((s) => s.id),
-    ).toEqual(['sub-a']);
-
-    const savedFeature = saveStateSpy.mock.calls[0][0].integrations.installed[0].features[0];
-    expect(savedFeature.subfeatures).toHaveLength(1);
-    expect(savedFeature.subfeatures![0]).toMatchObject({ featureId: 'sub-a' });
-  });
-
-  it('restores previously active container subfeatures from recorded state, excluding newly added ones', async () => {
-    const capturedContexts: IntegrationContext[] = [];
-    const now = '2026-01-01T00:00:00.000Z';
-
-    const state = makeState();
-    state.integrations.installed.push({
-      id: 'integration-id',
-      integrationId: 'test-integration',
-      installedByCliVersion: '0.9.0',
-      installedAt: now,
-      updatedByCliVersion: '0.9.0',
-      updatedAt: now,
-      features: [
-        {
-          featureId: 'container-feature',
-          scope: 'project',
-          targetRoot: tempDir,
-          installedByCliVersion: '0.9.0',
-          installedAt: now,
-          updatedByCliVersion: '0.9.0',
-          updatedAt: now,
-          dependencies: [],
-          resources: [],
-          operations: [],
-          subfeatures: [{ featureId: 'sub-a', dependencies: [] }],
-        },
-      ],
-    });
-    loadStateSpy.mockReturnValue(state);
-
-    const container: FeatureContainer = {
-      id: 'container-feature',
-      displayName: 'Container feature',
-      subfeatures: [
-        { id: 'sub-a', displayName: 'Sub A' },
-        { id: 'sub-b', displayName: 'Sub B' },
-        { id: 'sub-c', displayName: 'Sub C' }, // newly added, not in recorded state
-      ],
-      defaultInstallSubfeatureIds: ['sub-b'],
-      operations: [
-        {
-          id: 'test-op',
-          apply: (ctx) => {
-            capturedContexts.push(ctx);
-          },
-        },
-      ],
-    };
-    const registry = new IntegrationRegistry();
-    registry.register({
-      id: 'test-integration',
-      displayName: 'Test integration',
-      features: [container],
-    });
-
-    await migrateDeclarativeIntegrations(registry);
-
-    expect(capturedContexts).toHaveLength(1);
-    expect(
-      (capturedContexts[0] as ContainerIntegrationContext).activeSubfeatures.map((s) => s.id),
-    ).toEqual(['sub-a']);
-
-    const savedFeature = saveStateSpy.mock.calls[0][0].integrations.installed[0].features[0];
-    expect(savedFeature.subfeatures).toHaveLength(1);
-    expect(savedFeature.subfeatures![0]).toMatchObject({ featureId: 'sub-a' });
-  });
-
-  it('refreshes recorded subfeature resources and skips newly added subfeature resources', async () => {
-    const now = '2026-01-01T00:00:00.000Z';
-    const recordedPath = join(tempDir, 'recorded-sub.txt');
-    const newPath = join(tempDir, 'new-sub.txt');
-
-    const state = makeState();
-    state.integrations.installed.push({
-      id: 'integration-id',
-      integrationId: 'test-integration',
-      installedByCliVersion: '0.9.0',
-      installedAt: now,
-      updatedByCliVersion: '0.9.0',
-      updatedAt: now,
-      features: [
-        {
-          featureId: 'container-feature',
-          scope: 'project',
-          targetRoot: tempDir,
-          installedByCliVersion: '0.9.0',
-          installedAt: now,
-          updatedByCliVersion: '0.9.0',
-          updatedAt: now,
-          dependencies: [],
-          resources: [],
-          operations: [],
-          subfeatures: [{ featureId: 'sub-a', dependencies: [] }],
-        },
-      ],
-    });
-    loadStateSpy.mockReturnValue(state);
-
-    const container: FeatureContainer = {
-      id: 'container-feature',
-      displayName: 'Container feature',
-      subfeatures: [
-        {
-          id: 'sub-a',
-          displayName: 'Sub A',
-          resources: [wholeFile({ id: 'sub-a-file', targetPath: recordedPath, content: 'a' })],
-        },
-        // Never opted into, so update must not install it.
-        {
-          id: 'sub-b',
-          displayName: 'Sub B',
-          resources: [wholeFile({ id: 'sub-b-file', targetPath: newPath, content: 'b' })],
-        },
-      ],
-      defaultInstallSubfeatureIds: [],
-    };
-    const registry = new IntegrationRegistry();
-    registry.register({
-      id: 'test-integration',
-      displayName: 'Test integration',
-      features: [container],
-    });
-
-    await migrateDeclarativeIntegrations(registry);
-
-    expect(fs.readFileSync(recordedPath, 'utf-8')).toBe('a');
-    expect(fs.existsSync(newPath)).toBeFalse();
-
-    const savedFeature = saveStateSpy.mock.calls[0][0].integrations.installed[0].features[0];
-    expect(savedFeature.subfeatures![0].resources!.map((r) => r.id)).toEqual(['sub-a-file']);
-  });
-
-  it('applies plain feature normally when old state has container subfeatures recorded', async () => {
-    const capturedContexts: IntegrationContext[] = [];
-    const now = '2026-01-01T00:00:00.000Z';
-
-    const state = makeState();
-    state.integrations.installed.push({
-      id: 'integration-id',
-      integrationId: 'test-integration',
-      installedByCliVersion: '0.9.0',
-      installedAt: now,
-      updatedByCliVersion: '0.9.0',
-      updatedAt: now,
-      features: [
-        {
-          featureId: 'plain-feature',
-          scope: 'project',
-          targetRoot: tempDir,
-          installedByCliVersion: '0.9.0',
-          installedAt: now,
-          updatedByCliVersion: '0.9.0',
-          updatedAt: now,
-          dependencies: [],
-          resources: [],
-          operations: [],
-          subfeatures: [{ featureId: 'old-sub', dependencies: [] }],
-        },
-      ],
-    });
-    loadStateSpy.mockReturnValue(state);
-
-    const registry = new IntegrationRegistry();
-    registry.register({
-      id: 'test-integration',
-      displayName: 'Test integration',
-      features: [
-        {
-          id: 'plain-feature',
-          displayName: 'Plain feature',
-          operations: [
-            {
-              id: 'test-op',
-              apply: (ctx) => {
-                capturedContexts.push(ctx);
-              },
-            },
-          ],
-        },
-      ],
-    });
-
-    await migrateDeclarativeIntegrations(registry);
-
-    expect(capturedContexts).toHaveLength(1);
-    expect('activeSubfeatures' in capturedContexts[0]).toBeFalse();
-
-    const savedFeature = saveStateSpy.mock.calls[0][0].integrations.installed[0].features[0];
-    expect(savedFeature.subfeatures).toBeUndefined();
-  });
-
-  it('migrates replaced features into one successor with merged attrs', async () => {
-    const appliedAttrs: (IntegrationContext['attrs'] | undefined)[] = [];
-    const state = makeState();
-    state.integrations.installed.push({
-      id: 'integration-id',
-      integrationId: 'test-integration',
-      installedByCliVersion: '0.9.0',
-      installedAt: '2026-01-01T00:00:00.000Z',
-      updatedByCliVersion: '0.9.0',
-      updatedAt: '2026-01-01T00:00:00.000Z',
-      features: [
-        recordedFeature('old-sqaa', { projectKey: 'project-key' }),
-        recordedFeature('old-context', { orgKey: 'org-key', scaEnabled: true }),
-      ],
-    });
-    loadStateSpy.mockReturnValue(state);
-
-    const registry = new IntegrationRegistry();
-    registry.register({
-      id: 'test-integration',
-      displayName: 'Test integration',
-      features: [
-        {
-          id: 'vortex',
-          displayName: 'Vortex',
-          replacedIds: ['old-sqaa', 'old-context'],
-          operations: [
-            {
-              id: 'vortex-operation',
-              apply: (context) => {
-                appliedAttrs.push(context.attrs);
-              },
-            },
-          ],
-        },
-      ],
-    });
-
-    await migrateDeclarativeIntegrations(registry);
-
-    expect(appliedAttrs).toEqual([
-      { projectKey: 'project-key', orgKey: 'org-key', scaEnabled: true },
-    ]);
-    expect(state.integrations.installed[0].features).toHaveLength(1);
-    expect(state.integrations.installed[0].features[0]).toMatchObject({
-      featureId: 'vortex',
-      attrs: { projectKey: 'project-key', orgKey: 'org-key', scaEnabled: true },
-    });
-  });
-
-  it('merges predecessor attrs in successor replacement order', async () => {
-    const appliedAttrs: (IntegrationContext['attrs'] | undefined)[] = [];
-    const state = makeState();
-    state.integrations.installed.push({
-      id: 'integration-id',
-      integrationId: 'test-integration',
-      installedByCliVersion: '0.9.0',
-      installedAt: '2026-01-01T00:00:00.000Z',
-      updatedByCliVersion: '0.9.0',
-      updatedAt: '2026-01-01T00:00:00.000Z',
-      features: [
-        recordedFeature('old-context', { projectKey: 'context-project' }),
-        recordedFeature('old-sqaa', { projectKey: 'sqaa-project' }),
-      ],
-    });
-    loadStateSpy.mockReturnValue(state);
-
-    const registry = new IntegrationRegistry();
-    registry.register({
-      id: 'test-integration',
-      displayName: 'Test integration',
-      features: [
-        {
-          id: 'vortex',
-          displayName: 'Vortex',
-          replacedIds: ['old-sqaa', 'old-context'],
-          operations: [
-            {
-              id: 'vortex-operation',
-              apply: (context) => {
-                appliedAttrs.push(context.attrs);
-              },
-            },
-          ],
-        },
-      ],
-    });
-
-    await migrateDeclarativeIntegrations(registry);
-
-    expect(appliedAttrs).toEqual([{ projectKey: 'context-project' }]);
-  });
-
-  it('reapplies a recorded successor without merging predecessor attrs', async () => {
-    const appliedAttrs: (IntegrationContext['attrs'] | undefined)[] = [];
-    const state = makeState();
-    state.integrations.installed.push({
-      id: 'integration-id',
-      integrationId: 'test-integration',
-      installedByCliVersion: '0.9.0',
-      installedAt: '2026-01-01T00:00:00.000Z',
-      updatedByCliVersion: '0.9.0',
-      updatedAt: '2026-01-01T00:00:00.000Z',
-      features: [
-        recordedFeature('old-context', { orgKey: 'old-org' }),
-        recordedFeature('vortex', { projectKey: 'project-key' }),
-      ],
-    });
-    loadStateSpy.mockReturnValue(state);
-
-    const registry = new IntegrationRegistry();
-    registry.register({
-      id: 'test-integration',
-      displayName: 'Test integration',
-      features: [
-        {
-          id: 'vortex',
-          displayName: 'Vortex',
-          replacedIds: ['old-context'],
-          operations: [
-            {
-              id: 'vortex-operation',
-              apply: (context) => {
-                appliedAttrs.push(context.attrs);
-              },
-            },
-          ],
-        },
-      ],
-    });
-
-    await migrateDeclarativeIntegrations(registry);
-
-    expect(appliedAttrs).toEqual([{ projectKey: 'project-key' }]);
-    expect(state.integrations.installed[0].features).toHaveLength(1);
-    expect(state.integrations.installed[0].features[0].featureId).toBe('vortex');
   });
 });
